// apps/feedback-console/src/types/auth.ts
export interface CreateUserRequest {
  username?: string
  avatar?: string
}

/** TODO 后端用户信息 */
export interface TodoUser {
  id: number
  uuid: string
  username: string
  avatar: string
  created_at: string
  updated_at: string
}